\name{genes_human_h37}
\alias{genes_human_h37}
\docType{data}
\title{
Gene coordinates hg37 assembly.
}
\description{
Gene coordinates hg37 assembly.
}
\usage{data("genes_human_h37")}
\format{
  A data frame with 56638 observations on the following 4 variables.
  \describe{
    \item{\code{chr}}{a character vector}
    \item{\code{start}}{a numeric vector}
    \item{\code{end}}{a numeric vector}
    \item{\code{gene}}{a factor with levels ... }
  }
}
\details{
a
}
\source{
a
}
\references{
a
}
\examples{
data(genes_human_h37)
head(genes_human_h37)
}
\keyword{datasets}
