\name{interact.table}
\alias{interact.table}
\docType{data}
\title{
Gene name and number of interactions per gene.
}
\description{
Gene name and number of interactions per gene.
}
\usage{data("interact.table")}
\format{
  The format is:
 'table' int [1:8630(1d)] 2 8 4 5 18 8 15 3 1 4 ...
 - attr(*, "dimnames")=List of 1
  ..$ : chr [1:8630] "10'-apo-beta-carotenal" "10-formyltetrahydrofolic acid" "10-Formyl-THF-polyglutamate" "11,12-DHET" ...
}
\details{
a
}
\source{
a
}
\references{
a
}
\examples{
data(interact.table)
head(interact.table)
}
\keyword{datasets}
