\name{microCosm_v5_20}
\alias{microCosm_v5_20}
\docType{data}
\title{
a
}
\description{
a
}
\usage{data("microCosm_v5_20")}
\format{
  A data frame with 563179 observations on the following 12 variables.
  \describe{
    \item{\code{mir18}}{a character vector}
    \item{\code{mir17}}{a character vector}
    \item{\code{mir_acc}}{a character vector}
    \item{\code{mirmicrocosm}}{a character vector}
    \item{\code{target_name}}{a character vector}
    \item{\code{target_entrezid}}{a character vector}
    \item{\code{pval}}{a numeric vector}
    \item{\code{score}}{a numeric vector}
    \item{\code{method}}{a character vector}
    \item{\code{names}}{a character vector}
    \item{\code{mir20}}{a character vector}
    \item{\code{mir19}}{a character vector}
  }
}
\details{
a
}
\source{
a
}
\references{
a
}
\examples{
data(microCosm_v5_20)
head(microCosm_v5_20)
}
\keyword{datasets}
