\name{miRData-package}
\alias{miRData-package}
\alias{miRData}
\docType{package}
\title{
MiRNA data.
}
\description{
The miRData package stores some of the data needed by the \kbd{miRComb} package.
}
\details{
\tabular{ll}{
Package: \tab miRData\cr
Type: \tab Package\cr
Version: \tab 0.5.0\cr
Date: \tab 2015-10-05\cr
License: \tab GPL-3\cr
}

}
\author{
Maria Vila-Casadesús, Juanjo Lozano

Maintainer: Maria Vila-Casadesús <maria.vila@ciberehd.org>
}
\keyword{ package }
\seealso{
\kbd{miRComb} package
}
\examples{
library(miRData)
data(microCosm_v5_18)
data(targetScan_v6.2_18)
}
