\name{mirnas_human_18_h37}
\alias{mirnas_human_18_h37}
\docType{data}
\title{
MiRNA human (miRBase v18) coordinates
}
\description{
MiRNA human (miRBase v18) coordinates
}
\usage{data("mirnas_human_18_h37")}
\format{
  A data frame with 1919 observations on the following 4 variables.
  \describe{
    \item{\code{chr}}{a character vector}
    \item{\code{start}}{a numeric vector}
    \item{\code{end}}{a numeric vector}
    \item{\code{mirna}}{a factor with levels ...}
  }
}
\details{
a
}
\source{
a
}
\references{
a
}
\examples{
data(mirnas_human_18_h37)
head(mirnas_human_18_h37)
}
\keyword{datasets}
