\name{targetScan_v6.2_20}
\alias{targetScan_v6.2_20}
\docType{data}
\title{
TargetScan with mirbase v20 names
}
\description{
TargetScan with mirbase v20 names
}
\usage{data("targetScan_v6.2_20")}
\format{
  A data frame with 519097 observations on the following 10 variables.
  \describe{
    \item{\code{mir18}}{a character vector}
    \item{\code{mir17}}{a character vector}
    \item{\code{mir_acc}}{a character vector}
    \item{\code{mirtargetscan}}{a character vector}
    \item{\code{target_name}}{a character vector}
    \item{\code{target_entrezid}}{a character vector}
    \item{\code{method}}{a character vector}
    \item{\code{names}}{a character vector}
    \item{\code{mir20}}{a character vector}
    \item{\code{mir19}}{a character vector}
  }
}
\details{
a
}
\source{
a
}
\references{
a
}
\examples{
data(targetScan_v6.2_20)
head(targetScan_v6.2_20)
}
\keyword{datasets}
