\name{versions.mirnas}
\alias{versions.mirnas}
\docType{data}
\title{
MiRBase from 6.0 to 21
}
\description{
MiRBase from 6.0 to 21
}
\usage{data("versions.mirnas")}
\format{
  A data frame with 36339 observations on the following 23 variables.
  \describe{
   \item{\code{id}}{MIMAT unique identifier.}
   \item{\code{miRBase_6.0}}{version 6.0 of miRBase.}
   \item{\code{miRBase_7.0}}{version 7.0 of miRBase.}
   \item{\code{miRBase_7.1}}{version 7.1 of miRBase.}
   \item{\code{miRBase_8.0}}{version 8.0 of miRBase.}
   \item{\code{miRBase_8.1}}{version 8.1 of miRBase.}
   \item{\code{miRBase_8.2}}{version 8.2 of miRBase.}
   \item{\code{miRBase_9.0}}{version 9.0 of miRBase.}
   \item{\code{miRBase_9.1}}{version 9.1 of miRBase.}
   \item{\code{miRBase_9.2}}{version 9.2 of miRBase.}
   \item{\code{miRBase_10.0}}{version 10.0 of miRBase.}
   \item{\code{miRBase_10.1}}{version 10.1 of miRBase.}
   \item{\code{miRBase_11.0}}{version 11.0 of miRBase.}
   \item{\code{miRBase_12.0}}{version 12.0 of miRBase.}
   \item{\code{miRBase_13.0}}{version 13.0 of miRBase.}
   \item{\code{miRBase_14}}{version 14 of miRBase.}
   \item{\code{miRBase_15}}{version 15 of miRBase.}
   \item{\code{miRBase_16}}{version 16 of miRBase.}
   \item{\code{miRBase_17}}{version 17 of miRBase.}
   \item{\code{miRBase_18}}{version 18 of miRBase.}
   \item{\code{miRBase_19}}{version 19 of miRBase.}
   \item{\code{miRBase_20}}{version 20 of miRBase.}
   \item{\code{miRBase_21}}{version 21 of miRBase.}
}
}
\details{
}
\source{
\url{www.mirbase.org}
}
\references{

}
\examples{
data(versions.mirnas)
## maybe str(versions.mirnas) ; plot(versions.mirnas) ...
}
\keyword{datasets}
